Move logic tests file inside logic folder
